Bug# 1614 Removed the FLOAT type from EtomoNumber; storing all floating point values as doubles.

+++ Etomo/src/etomo/ApplicationManager.java
@@ -6022,7 +6022,7 @@ public final class ApplicationManager extends BaseManager implements
     comScriptMgr.loadVolcombine();
     // try to load reduction factor
     ConstSetParam setParam = comScriptMgr.getSetParamFromVolcombine(
-        SetParam.COMBINEFFT_REDUCTION_FACTOR_NAME, EtomoNumber.Type.FLOAT);
+        SetParam.COMBINEFFT_REDUCTION_FACTOR_NAME, EtomoNumber.Type.DOUBLE);
     tomogramCombinationDialog.setReductionFactorParams(setParam);
     tomogramCombinationDialog.enableReductionFactor(setParam != null
         && setParam.isValid());
@@ -6083,7 +6083,7 @@ public final class ApplicationManager extends BaseManager implements
       // Make sure the reduction factor set command is available in
       // volcombine.com
       SetParam setParam = comScriptMgr.getSetParamFromVolcombine(
-          SetParam.COMBINEFFT_REDUCTION_FACTOR_NAME, EtomoNumber.Type.FLOAT);
+          SetParam.COMBINEFFT_REDUCTION_FACTOR_NAME, EtomoNumber.Type.DOUBLE);
       boolean setParamIsValid = setParam != null && setParam.isValid();
       tomogramCombinationDialog.enableReductionFactor(setParamIsValid);
       tomogramCombinationDialog.getReductionFactorParam(setParam);
